# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [idea, building, live, paused, abandoned, flopped]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes and ideas
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /keywords:
    get:
      operationId: listKeywords
      tags: [keywords]
      summary: List keyword research entries
      parameters:
        - name: niche
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [researching, targeting, ranking, dropped]
      responses:
        "200":
          description: List of keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: createKeyword
      tags: [keywords]
      summary: Add a keyword research entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeywordBody"
      responses:
        "201":
          description: Created keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
  /keywords/{id}:
    patch:
      operationId: updateKeyword
      tags: [keywords]
      summary: Update a keyword entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKeywordBody"
      responses:
        "200":
          description: Updated keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
    delete:
      operationId: deleteKeyword
      tags: [keywords]
      summary: Delete a keyword entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /planner:
    get:
      operationId: listPlannerItems
      tags: [planner]
      summary: List planner items
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [planned, in_progress, done, skipped]
      responses:
        "200":
          description: List of planner items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlannerItem"
    post:
      operationId: createPlannerItem
      tags: [planner]
      summary: Create a planner item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlannerBody"
      responses:
        "201":
          description: Created planner item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerItem"
  /planner/{id}:
    patch:
      operationId: updatePlannerItem
      tags: [planner]
      summary: Update a planner item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlannerBody"
      responses:
        "200":
          description: Updated planner item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerItem"
    delete:
      operationId: deletePlannerItem
      tags: [planner]
      summary: Delete a planner item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [idea, building, live, paused, abandoned, flopped]
        url:
          type: ["string", "null"]
        traffic:
          type: integer
        revenue:
          type: number
        nextTask:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, type, status, traffic, revenue, createdAt, updatedAt]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [idea, building, live, paused, abandoned, flopped]
          default: idea
        url:
          type: ["string", "null"]
        traffic:
          type: integer
          default: 0
        revenue:
          type: number
          default: 0
        nextTask:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required: [name, type]
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [idea, building, live, paused, abandoned, flopped]
        url:
          type: ["string", "null"]
        traffic:
          type: integer
        revenue:
          type: number
        nextTask:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: ["string", "null"]
        category:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, category, pinned, createdAt, updatedAt]
    CreateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: ["string", "null"]
        category:
          type: string
          default: general
        pinned:
          type: boolean
          default: false
      required: [title]
    UpdateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: ["string", "null"]
        category:
          type: string
        pinned:
          type: boolean
    Keyword:
      type: object
      properties:
        id:
          type: integer
        keyword:
          type: string
        niche:
          type: string
        volume:
          type: integer
        difficulty:
          type: integer
        cpc:
          type: number
        competition:
          type: string
          enum: [low, medium, high]
        position:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [researching, targeting, ranking, dropped]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, keyword, niche, volume, difficulty, cpc, competition, status, createdAt, updatedAt]
    CreateKeywordBody:
      type: object
      properties:
        keyword:
          type: string
        niche:
          type: string
        volume:
          type: integer
          default: 0
        difficulty:
          type: integer
          default: 0
        cpc:
          type: number
          default: 0
        competition:
          type: string
          enum: [low, medium, high]
          default: low
        position:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [researching, targeting, ranking, dropped]
          default: researching
      required: [keyword, niche]
    UpdateKeywordBody:
      type: object
      properties:
        keyword:
          type: string
        niche:
          type: string
        volume:
          type: integer
        difficulty:
          type: integer
        cpc:
          type: number
        competition:
          type: string
          enum: [low, medium, high]
        position:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [researching, targeting, ranking, dropped]
...
    PlannerItem:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        projectId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planned, in_progress, done, skipped]
        timeSlot:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, date, title, status, priority, createdAt, updatedAt]
    CreatePlannerBody:
      type: object
      properties:
        date:
          type: string
        projectId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planned, in_progress, done, skipped]
          default: planned
        timeSlot:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
      required: [date, title]
    UpdatePlannerBody:
      type: object
      properties:
        date:
          type: string
        projectId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planned, in_progress, done, skipped]
        timeSlot:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
...